+++ Data/SBV/Dynamic.hs
@@ -63,6 +63,7 @@ module Data.SBV.Dynamic
   , svToWord1, svFromWord1, svTestBit, svSetBit
   , svShiftLeft, svShiftRight
   , svRotateLeft, svRotateRight
+  , svBarrelRotateLeft, svBarrelRotateRight
   , svWordFromBE, svWordFromLE
   , svBlastLE, svBlastBE
   -- ** Conditionals: Mergeable values